+++ src/main/java/ontos/infovis/pojo/Composition.java
@@ -17,8 +17,8 @@ public class Composition {
   private String owner;
   private List<Right> rights;
 
-  private int creation_date;
-  private int last_update;
+  private long creation_date;
+  private long last_update;
   private String structure;
   private List<ComponentInstance> components;
 
@@ -81,7 +81,7 @@ public class Composition {
    *
    * @param last_update New value of last_update.
    */
-  public void setLast_update(int last_update) {
+  public void setLast_update(long last_update) {
     this.last_update = last_update;
   }
 
@@ -90,7 +90,7 @@ public class Composition {
    *
    * @param creation_date New value of creation_date.
    */
-  public void setCreation_date(int creation_date) {
+  public void setCreation_date(long creation_date) {
     this.creation_date = creation_date;
   }
 
@@ -108,7 +108,7 @@ public class Composition {
    *
    * @return Value of last_update.
    */
-  public int getLast_update() {
+  public long getLast_update() {
     return last_update;
   }
 
@@ -144,7 +144,7 @@ public class Composition {
    *
    * @return Value of creation_date.
    */
-  public int getCreation_date() {
+  public long getCreation_date() {
     return creation_date;
   }
